Fuxing raise condition error during the generation python files for parsing XDL files.
This situation occurs when the same type of XDL files are parsed in
different threads, which leads to the simultaneous generation of
python files used for parsing. In this case, the first thread has
generated these python files, begins to use them, and at the same
time, the second thread then generates the same python files,
thereby deleting the existing ones and refilling their contents.
At this moment the conflict occurs.

To avoid this situation new function 'generate_xdl_ext' was
implemented and used in 'generate_xdls'.

Old function 'generate_xdls' is also presents in the code.

Signed-off-by: Dmitriy Terletskiy <daterletskiy@gmail.com>