Fixed issue with setting `ROOT_SOURCE_DIR` variable.
Before this value was equal to `CMAKE_SOURCE_DIR` because all
sources were placed inside current builder directory.

Since building sources out of the tree is supported this value
is not relevant and must be substituted with `SOURCE_DIR` variable
value what are passed as the parameter to cmake during the
configuration.

Signed-off-by: Dmitriy Terletskiy <daterletskiy@gmail.com>